@@ -18,78 +18,78 @@
     "test:watch": "npm run lint && jest --coverage --watch"
   },
   "dependencies": {
-    "@loadable/component": "^5.12.0",
-    "@mdx-js/mdx": "^1.5.9",
-    "@mdx-js/react": "^1.5.9",
-    "axios": "^0.19.2",
-    "babel-plugin-styled-components": "^1.10.7",
-    "gatsby": "^2.20.28",
-    "gatsby-image": "^2.3.4",
-    "gatsby-plugin-manifest": "^2.3.6",
-    "gatsby-plugin-mdx": "^1.1.11",
-    "gatsby-plugin-offline": "^3.1.4",
-    "gatsby-plugin-postcss": "^2.2.3",
-    "gatsby-plugin-react-helmet": "^3.2.4",
-    "gatsby-plugin-sharp": "^2.5.6",
-    "gatsby-plugin-sitemap": "^2.3.5",
-    "gatsby-plugin-styled-components": "^3.2.4",
-    "gatsby-plugin-svgr": "^2.0.2",
-    "gatsby-plugin-webpack-size": "^1.0.0",
-    "gatsby-source-filesystem": "^2.2.4",
-    "gatsby-transformer-json": "^2.3.3",
-    "gatsby-transformer-sharp": "^2.4.6",
-    "gatsby-transformer-yaml": "^2.3.3",
+    "@loadable/component": "^5.16.7",
+    "@mdx-js/mdx": "^3.1.0",
+    "@mdx-js/react": "^3.1.0",
+    "axios": "^1.10.0",
+    "babel-plugin-styled-components": "^2.1.4",
+    "framer-motion": "^12.23.6",
+    "gatsby": "^5.14.5",
+    "gatsby-image": "^3.11.0",
+    "gatsby-plugin-manifest": "^5.14.0",
+    "gatsby-plugin-mdx": "^5.14.1",
+    "gatsby-plugin-offline": "^6.14.0",
+    "gatsby-plugin-postcss": "^6.14.0",
+    "gatsby-plugin-react-helmet": "^6.14.0",
+    "gatsby-plugin-sharp": "^5.14.0",
+    "gatsby-plugin-sitemap": "^6.14.0",
+    "gatsby-plugin-styled-components": "^6.14.0",
+    "gatsby-plugin-svgr": "^3.0.0-beta.0",
+    "gatsby-plugin-webpack-size": "^2.0.1",
+    "gatsby-source-filesystem": "^5.14.0",
+    "gatsby-transformer-json": "^5.14.0",
+    "gatsby-transformer-sharp": "^5.14.0",
+    "gatsby-transformer-yaml": "^5.14.0",
     "giphy-js-sdk-core": "^1.0.6",
-    "graphql": "^14.6.0",
-    "intersection-observer": "^0.9.0",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-feather": "^2.0.4",
-    "react-helmet": "^6.0.0",
-    "react-pose": "^4.0.10",
-    "remark": "^12.0.0",
-    "remark-breaks": "^1.0.5",
-    "remark-html": "^11.0.0",
-    "remark-parse": "^8.0.1",
-    "remark-react": "^7.0.0",
-    "shortid": "^2.2.15",
-    "styled-components": "^5.1.0",
+    "graphql": "^16.11.0",
+    "intersection-observer": "^0.12.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-feather": "^2.0.10",
+    "react-helmet": "^6.1.0",
+    "remark": "^15.0.1",
+    "remark-breaks": "^4.0.0",
+    "remark-html": "^16.0.1",
+    "remark-parse": "^11.0.0",
+    "remark-react": "^10.0.0",
+    "shortid": "^2.2.17",
+    "styled-components": "^6.1.19",
     "use-dark-mode": "^2.3.1"
   },
   "devDependencies": {
-    "@babel/core": "^7.9.0",
-    "@babel/node": "^7.8.7",
-    "@babel/preset-env": "^7.9.5",
+    "@babel/core": "^7.28.0",
+    "@babel/eslint-parser": "^7.28.0",
+    "@babel/node": "^7.28.0",
+    "@babel/preset-env": "^7.28.0",
     "@react-mock/localstorage": "^0.1.2",
-    "@svgr/webpack": "^5.3.1",
-    "@testing-library/jest-dom": "^5.5.0",
-    "@testing-library/react": "^10.0.3",
-    "@welldone-software/why-did-you-render": "^4.1.1",
-    "babel-eslint": "^10.1.0",
-    "babel-jest": "^25.4.0",
-    "babel-preset-gatsby": "^0.3.3",
-    "chalk": "^4.0.0",
-    "eslint": "^6.8.0",
-    "eslint-config-prettier": "^6.11.0",
-    "eslint-loader": "^4.0.0",
-    "eslint-plugin-graphql": "^3.1.1",
-    "eslint-plugin-prettier": "^3.1.3",
-    "eslint-plugin-react": "^7.19.0",
-    "eslint-plugin-react-hooks": "^3.0.0",
-    "eslint-plugin-testing-library": "^3.0.4",
-    "gatsby-plugin-webpack-bundle-analyser-v2": "^1.1.8",
+    "@svgr/webpack": "^8.1.0",
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/react": "^16.3.0",
+    "@welldone-software/why-did-you-render": "^10.0.1",
+    "babel-jest": "^30.0.4",
+    "babel-preset-gatsby": "^3.14.0",
+    "chalk": "^5.4.1",
+    "eslint": "^9.31.0",
+    "eslint-config-prettier": "^10.1.8",
+    "eslint-loader": "^4.0.2",
+    "eslint-plugin-graphql": "^4.0.0",
+    "eslint-plugin-prettier": "^5.5.3",
+    "eslint-plugin-react": "^7.37.5",
+    "eslint-plugin-react-hooks": "^5.2.0",
+    "eslint-plugin-testing-library": "^7.6.0",
+    "gatsby-plugin-webpack-bundle-analyser-v2": "^1.1.32",
     "identity-obj-proxy": "^3.0.0",
-    "jest": "^25.4.0",
-    "jest-canvas-mock": "^2.2.0",
-    "js-yaml": "^3.13.1",
-    "ora": "^4.0.4",
-    "postcss-preset-env": "^6.7.0",
+    "jest": "^30.0.4",
+    "jest-canvas-mock": "^2.5.2",
+    "js-yaml": "^4.1.0",
+    "ora": "^8.2.0",
+    "postcss-preset-env": "^10.2.4",
     "prepend": "^1.0.2",
-    "prettier": "^2.0.5",
-    "slugify": "^1.4.0",
-    "stylelint": "^13.3.3",
-    "stylelint-config-prettier": "^8.0.1",
-    "stylelint-prettier": "^1.1.2"
+    "prettier": "^3.6.2",
+    "slugify": "^1.6.6",
+    "stylelint": "^16.22.0",
+    "stylelint-config-prettier": "^9.0.5",
+    "stylelint-prettier": "^5.0.3"
   },
   "browserslist": [
     ">0.2%",
