@@ -1,6 +1,5 @@
-import React from 'react'
 import PropTypes from 'prop-types'
-import styles from './Button.module.css'
+import * as styles from './Button.module.css'
 
 const Button = (props) => (
   <a className={styles.button} {...props}>
