@@ -1,6 +1,5 @@
-import React from 'react'
 import PropTypes from 'prop-types'
-import styles from './FullWidth.module.css'
+import * as styles from './FullWidth.module.css'
 
 const FullWidth = ({ children }) => (
   <div className={styles.fullWidth}>{children}</div>
