@@ -1,8 +1,8 @@
-import React, { PureComponent } from 'react'
-import PropTypes from 'prop-types'
 import { graphql } from 'gatsby'
 import Img from 'gatsby-image'
-import styles from './ProjectImage.module.css'
+import PropTypes from 'prop-types'
+import { PureComponent } from 'react'
+import * as styles from './ProjectImage.module.css'
 
 export default class ProjectImage extends PureComponent {
   static propTypes = {
