@@ -1,17 +1,36 @@
-import React, { useEffect, useRef, useState } from 'react'
-import posed from 'react-pose'
-import { Link } from 'gatsby'
-import styles from './Navbar.module.css'
-import { moveInBottom } from '../atoms/Transitions'
-import Logo from '../../images/AppIcon.png'
-import { useApp } from '../../hooks/use-app'
+// src/components/organisms/Navbar.jsx
 
-const activeClassName = 'active'
+import { motion } from 'framer-motion' // 'posed' yerine 'motion' import ediyoruz
+import { Link } from 'gatsby'
+import React, { useRef, useState } from 'react' // useEffect vs. kullanıldığı için React import'u GEREKLİ!
+import { useApp } from '../../hooks/use-app'
+import Logo from '../../images/AppIcon.png'
+import * as styles from './Navbar.module.css'
+
+const activeClassName = styles.active // 'active' string'i yerine CSS modülünden gelen sınıfı kullanmak daha güvenlidir
 
 const Navbar = () => {
   const { basics } = useApp()
-  const Animation = posed.div(moveInBottom)
 
+  // 'react-pose' yerine 'framer-motion' için animasyon varyantları
+  // Bu, eski `moveInBottom` geçişinizin karşılığıdır.
+  const navbarVariants = {
+    hidden: {
+      opacity: 0,
+      y: -20 // Yukarıdan gelsin
+    },
+    visible: {
+      opacity: 1,
+      y: 0,
+      transition: {
+        delay: 0.2, // Hafif bir gecikme
+        duration: 0.5,
+        ease: 'easeInOut'
+      }
+    }
+  }
+
+  // Bu bölüm (isSticky logiği) aynı kalıyor, zaten doğru çalışıyor.
   const [isSticky, setSticky] = useState(false)
   const ref = useRef(null)
   const handleScroll = () => {
@@ -20,23 +39,32 @@ const Navbar = () => {
     }
   }
 
-  useEffect(() => {
+  React.useEffect(() => {
     window.addEventListener('scroll', handleScroll)
 
-
+    // DÜZELTME: removeEventListener içindeki fonksiyonun referansı
+    // addEventListener'daki ile aynı olmalı.
     return () => {
-      window.removeEventListener('scroll', () => handleScroll)
+      window.removeEventListener('scroll', handleScroll) // <-- DÜZELTİLDİ
     }
-  }, [])
+  }, []) // Boş dizi doğru, sadece component mount olduğunda çalışır.
 
   return (
-    <Animation className={styles.navbar}>
+    // 'posed.div' yerine 'motion.header' veya 'motion.nav' kullanmak
+    // anlamsal olarak daha doğrudur.
+    <motion.header
+      ref={ref}
+      className={`${styles.navbar} ${isSticky ? styles.sticky : ''}`}
+      variants={navbarVariants}
+      initial="hidden" // Başlangıç durumu
+      animate="visible" // Ekrana geldiğindeki durum
+    >
       <Link
         to="/"
         className={styles.logolink}
         activeClassName={activeClassName}
       >
-        <img className={styles.logo} src={Logo} />
+        <img className={styles.logo} src={Logo} alt={basics.label} />
         <span className={styles.logotitle}>{basics.label.toLowerCase()}</span>
       </Link>
       <div className={styles.menu}>
@@ -52,12 +80,15 @@ const Navbar = () => {
         <Link to="/changelog/" activeClassName={activeClassName}>
           Changelog
         </Link>
-        <Link to={basics.storeUrl} activeClassName={activeClassName} className={styles.ctaButton}>
+        <Link
+          to={basics.storeUrl}
+          activeClassName={activeClassName}
+          className={styles.ctaButton}
+        >
           {basics.ctaText}
         </Link>
-        
       </div>
-    </Animation>
+    </motion.header>
   )
 }
 
